requirements.txt
pybit>=5.15.0
deribit-api>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.1.0
aiohttp>=3.9.0
pydantic>=2.6.0
python-dotenv>=1.0.0
websockets>=12.0
cryptography>=42.0.0
aiofiles>=24.1.0
loguru>=0.7.0
